# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BookHub API
  description: |
    BookHub の Web API 仕様。Chrome 拡張機能からスクレイピングデータを受信・保存するエンドポイントを提供します。

    ## 認証

    ほとんどのエンドポイントは **Supabase の Bearer トークン**（ID トークン）による認証が必要です。
    トークンは Supabase Auth から取得します。

    ```
    Authorization: Bearer <SUPABASE_ID_TOKEN>
    ```

    ## Edge Runtime

    全エンドポイントは Cloudflare Pages の Edge Runtime で動作します。Node.js API の一部（`fs`, `crypto` 等）は使用不可です。
  version: 0.1.0
  contact:
    name: BookHub
servers:
  - url: https://bookhub.pages.dev
    description: 本番環境（Cloudflare Pages）
  - url: http://localhost:3000
    description: ローカル開発環境

tags:
  - name: scrape
    description: Chrome 拡張機能からのスクレイピングデータ受信
  - name: books
    description: 蔵書の CRUD 操作（Web アプリ用）

security:
  - bearerAuth: []

paths:
  /api/scrape:
    post:
      summary: スクレイピングデータ送信
      description: |
        Chrome 拡張機能が電子書籍ストア（Kindle・DMM 等）からスクレイピングした書籍データを受信し、Supabase に保存します。

        **処理の流れ:**
        1. Bearer トークンで認証（Supabase ID トークン）
        2. リクエストボディを Zod スキーマでバリデーション
        3. books テーブルで重複チェック（title + author + volumeNumber の組み合わせ）
        4. 新規書籍は `books` テーブルに INSERT
        5. `user_books` テーブルに UPSERT（重複送信は冪等）
        6. 異なるストアで既に所持している場合は `duplicates` に追加して通知

        **注意事項:**
        - `books` 配列は 1 件以上 500 件以下
        - リクエスト内に同一書籍が複数含まれる場合は自動的に重複排除
        - title / author の前後空白は自動除去（trim）
        - レート制限は Cloudflare WAF のルールで管理
      operationId: postScrape
      tags:
        - scrape
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrapePayload'
            examples:
              single_book:
                summary: 単一書籍（巻数あり）
                value:
                  books:
                    - title: ワンピース
                      author: 尾田栄一郎
                      volumeNumber: 107
                      store: kindle
                      isAdult: false
              single_volume:
                summary: 単巻作品（巻数なし）
                value:
                  books:
                    - title: 火花
                      author: 又吉直樹
                      store: kindle
                      isAdult: false
              adult_content:
                summary: 成人向けコンテンツ
                value:
                  books:
                    - title: テスト作品
                      author: テスト著者
                      store: dmm
                      isAdult: true
              multiple_books:
                summary: 複数書籍
                value:
                  books:
                    - title: ワンピース
                      author: 尾田栄一郎
                      volumeNumber: 106
                      store: kindle
                      isAdult: false
                    - title: ワンピース
                      author: 尾田栄一郎
                      volumeNumber: 107
                      store: kindle
                      isAdult: false
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrapeResponse'
              examples:
                no_duplicates:
                  summary: 重複なし
                  value:
                    savedCount: 2
                    duplicateCount: 0
                    duplicates: []
                with_duplicates:
                  summary: 重複あり（別ストアで既所持）
                  value:
                    savedCount: 1
                    duplicateCount: 1
                    duplicates:
                      - title: ワンピース
                        volumeNumber: 107
                        existingStores:
                          - dmm
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_json:
                  summary: 無効な JSON
                  value:
                    error: invalid_json
                    message: Request body is not valid JSON
                validation_error:
                  summary: バリデーション失敗
                  value:
                    error: validation_error
                    message: Request body validation failed
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_token:
                  summary: Authorization ヘッダーなし
                  value:
                    error: unauthorized
                    message: Missing Bearer token
                invalid_token:
                  summary: 無効・期限切れトークン
                  value:
                    error: unauthorized
                    message: Invalid or expired token
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  value:
                    error: internal_error
                    message: An unexpected error occurred

    options:
      summary: CORS プリフライト
      description: |
        Chrome 拡張機能向けの CORS プリフライトレスポンス。
        Chrome 拡張機能は CORS 制約を受けないため、`Access-Control-Allow-Origin` ヘッダーは返しません。
        ブラウザからの直接アクセスを防ぐための最小限のレスポンスです。
      operationId: optionsScrape
      tags:
        - scrape
      security: []
      responses:
        '204':
          description: プリフライト応答
          headers:
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: POST, OPTIONS
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: Authorization, Content-Type

  /api/books:
    get:
      summary: 蔵書一覧取得
      description: |
        ログインユーザーの蔵書一覧を取得します。タイトル/著者名での検索、ストアフィルタ、成人向けフィルタ、ページネーションに対応。
      operationId: getBooks
      tags:
        - books
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            minLength: 2
          description: タイトルまたは著者名の部分一致検索（2文字以上）
        - name: store
          in: query
          schema:
            $ref: '#/components/schemas/Store'
          description: ストアでフィルタ
        - name: isAdult
          in: query
          schema:
            type: string
            enum: ['true', 'false']
          description: 成人向けフィルタ
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: ページ番号
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 1ページあたりの件数
      responses:
        '200':
          description: 蔵書一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBooksResponse'
              examples:
                with_books:
                  summary: 蔵書あり
                  value:
                    books:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        title: ワンピース
                        author: 尾田栄一郎
                        volumeNumber: 107
                        thumbnailUrl: https://example.com/cover.jpg
                        isbn: '9784088835099'
                        publishedAt: '2024-03-04'
                        isAdult: false
                        createdAt: '2024-01-01T00:00:00Z'
                        userBookId: 660e8400-e29b-41d4-a716-446655440000
                        store: kindle
                        userBookCreatedAt: '2024-03-04T00:00:00Z'
                    total: 1
                    page: 1
                    limit: 20
                empty:
                  summary: 蔵書なし
                  value:
                    books: []
                    total: 0
                    page: 1
                    limit: 20
        '400':
          description: クエリパラメータ不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: 書籍を手動登録
      description: |
        手動で書籍を登録します。books マスタへの追加（既存の場合は再利用）と user_books への INSERT を行います。

        別ストアで既に所持している場合は `alreadyOwned: true` と `existingStores` で二度買い情報を返します。
        同一ストアで既に登録済みの場合は 409 Conflict を返します。
      operationId: registerBook
      tags:
        - books
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterBook'
            examples:
              minimal:
                summary: 必須フィールドのみ
                value:
                  title: ワンピース
                  author: 尾田栄一郎
                  store: kindle
              full:
                summary: 全フィールド指定
                value:
                  title: ワンピース
                  author: 尾田栄一郎
                  volumeNumber: 107
                  store: kindle
                  thumbnailUrl: https://example.com/cover.jpg
                  isbn: '9784088835099'
                  publishedAt: '2024-03-04'
                  isAdult: false
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterBookResponse'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 同一ストアで既に登録済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conflict:
                  value:
                    error: conflict
                    message: この書籍は既に kindle で登録されています
        '500':
          $ref: '#/components/responses/InternalError'

  /api/books/{id}:
    patch:
      summary: 所持書籍の store を更新
      description: |
        user_books レコードの store を更新します。`{id}` は user_books.id（UUID）です。
      operationId: updateUserBook
      tags:
        - books
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: user_books レコード ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserBook'
            examples:
              change_store:
                value:
                  store: dmm
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookWithStore'
        '400':
          description: UUID 不正またはバリデーション失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: 所持書籍を削除
      description: |
        user_books レコードを削除します。books マスタは削除しません（他ユーザーも参照している可能性があるため）。
        `{id}` は user_books.id（UUID）です。
      operationId: deleteUserBook
      tags:
        - books
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: user_books レコード ID
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Deleted
        '400':
          description: UUID 不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase の ID トークン（`supabase.auth.getSession()` で取得）

  schemas:
    Store:
      type: string
      enum:
        - kindle
        - dmm
        - other
      description: 電子書籍ストア識別子
      example: kindle

    ScrapeBook:
      type: object
      required:
        - title
        - author
        - store
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: 書籍タイトル（前後の空白は自動除去）
          example: ワンピース
        author:
          type: string
          minLength: 1
          maxLength: 200
          description: 著者名（前後の空白は自動除去）
          example: 尾田栄一郎
        volumeNumber:
          type: integer
          minimum: 1
          maximum: 9999
          description: 巻数（単巻作品の場合は省略可）
          example: 107
        store:
          $ref: '#/components/schemas/Store'
        thumbnailUrl:
          type: string
          format: uri
          description: 書影 URL（https のみ許可）
          example: https://example.com/thumbnail.jpg
        isbn:
          type: string
          pattern: '^\d{10}(\d{3})?$'
          description: ISBN-10 または ISBN-13（数字のみ）
          example: '9784088830490'
        isAdult:
          type: boolean
          default: false
          description: 成人向けコンテンツフラグ（本棚での表示制御に使用）
          example: false

    ScrapePayload:
      type: object
      required:
        - books
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/ScrapeBook'
          minItems: 1
          maxItems: 500
          description: 送信する書籍データのリスト

    ScrapeDuplicate:
      type: object
      required:
        - title
        - existingStores
      properties:
        title:
          type: string
          description: 重複が検知された書籍のタイトル
          example: ワンピース
        volumeNumber:
          type: integer
          description: 巻数（単巻作品の場合は省略）
          example: 107
        existingStores:
          type: array
          items:
            $ref: '#/components/schemas/Store'
          minItems: 1
          maxItems: 3
          uniqueItems: true
          description: 既に所持しているストアの一覧
          example:
            - dmm

    ScrapeResponse:
      type: object
      required:
        - savedCount
        - duplicateCount
        - duplicates
      properties:
        savedCount:
          type: integer
          minimum: 0
          description: 新規に保存された書籍数（既存の再送信も含む）
          example: 2
        duplicateCount:
          type: integer
          minimum: 0
          description: 異なるストアで既に所持していた書籍数（二度買い防止アラート対象）
          example: 1
        duplicates:
          type: array
          items:
            $ref: '#/components/schemas/ScrapeDuplicate'
          description: 重複が検知された書籍の詳細リスト

    RegisterBook:
      type: object
      required:
        - title
        - author
        - store
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: 書籍タイトル
          example: ワンピース
        author:
          type: string
          minLength: 1
          maxLength: 200
          description: 著者名
          example: 尾田栄一郎
        volumeNumber:
          type: integer
          minimum: 1
          maximum: 9999
          description: 巻数（単巻作品の場合は省略可）
          example: 107
        store:
          $ref: '#/components/schemas/Store'
        thumbnailUrl:
          type: string
          format: uri
          description: 書影 URL（https のみ）
        isbn:
          type: string
          pattern: '^\d{10}(\d{3})?$'
          description: ISBN-10 または ISBN-13
        publishedAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: 出版日（YYYY-MM-DD）
        isAdult:
          type: boolean
          default: false
          description: 成人向けフラグ

    UpdateUserBook:
      type: object
      required:
        - store
      properties:
        store:
          $ref: '#/components/schemas/Store'

    BookWithStore:
      type: object
      required:
        - id
        - title
        - author
        - isAdult
        - createdAt
        - userBookId
        - store
        - userBookCreatedAt
      properties:
        id:
          type: string
          format: uuid
          description: books テーブルの ID
        title:
          type: string
        author:
          type: string
        volumeNumber:
          type: integer
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        isAdult:
          type: boolean
        createdAt:
          type: string
          format: date-time
        userBookId:
          type: string
          format: uuid
          description: user_books テーブルの ID
        store:
          $ref: '#/components/schemas/Store'
        userBookCreatedAt:
          type: string
          format: date-time

    GetBooksResponse:
      type: object
      required:
        - books
        - total
        - page
        - limit
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookWithStore'
        total:
          type: integer
          minimum: 0
          description: 総件数
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1

    RegisterBookResponse:
      type: object
      required:
        - book
        - alreadyOwned
        - existingStores
      properties:
        book:
          $ref: '#/components/schemas/BookWithStore'
        alreadyOwned:
          type: boolean
          description: 別ストアで既に所持しているか
        existingStores:
          type: array
          items:
            $ref: '#/components/schemas/Store'
          description: 既に所持しているストアの一覧（二度買い防止情報）

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - unauthorized
            - invalid_json
            - validation_error
            - invalid_uuid
            - not_found
            - conflict
            - internal_error
          description: エラーコード（機械的なハンドリング用）
          example: unauthorized
        message:
          type: string
          description: 人間が読めるエラー説明
          example: Missing Bearer token

  responses:
    Unauthorized:
      description: 認証失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              value:
                error: unauthorized
                message: Missing Bearer token
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              value:
                error: not_found
                message: 指定されたレコードが見つかりません
    InternalError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              value:
                error: internal_error
                message: An unexpected error occurred